skills/md2epub/pyproject.toml
[project]
name = "md2epub"
version = "1.0.0"
description = "Convert Markdown to EPUB for e-readers"
requires-python = ">=3.9"
dependencies = [
    "ebooklib>=0.18",
    "markdown>=3.4",
    "Pillow>=9.0",
    "requests>=2.28",
    "PyYAML>=6.0",
    "certifi>=2023.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
